+++ core/src/test/scala/com/salesforce/op/stages/impl/feature/MultiPickListMapVectorizerTest.scala
@@ -544,7 +544,7 @@ class MultiPickListMapVectorizerTest extends FlatSpec with TestSparkContext with
   }
 
   it should "drop features with max cardinality" in {
-    val fitted = vectorizer.setMaxPercentageCardinality(0.01)
+    val fitted = vectorizer.setMaxPctCardinality(0.01)
       .fit(dataSet)
     val transformed = fitted.transform(dataSet)
     val vectorMetadata = fitted.getMetadata()
